Brought back the original readme file.
As I'm learning the ropes with GitBook, I had some troubles where GitBook erased the content from the readme as I was setting up the synchronization with the repository. 

It seems to be generating a new markdown file per new page now that I'm creating so things should be ok now.